将搜索Activity部分内容迁移到搜索Fragment
