+++ app/src/main/res/values/strings.xml
@@ -9,5 +9,8 @@
     <string name="password">password</string>
     <string name="login">login</string>
 
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
+
 
 </resources>
